Fix compilation warning (about format mismatch)
git-svn-id: http://root.cern.ch/svn/root/trunk@36573 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/cint/src/common.h
@@ -1304,7 +1304,7 @@ public:
    void Print() {
       NameMap_t::iterator iMap = fMap.begin();
       while( iMap != fMap.end() ) {
-         fprintf(stderr,"key=%s size=%ld\n",iMap->first,iMap->second.size());
+         fprintf(stderr,"key=%s size=%ld\n",iMap->first,(long)iMap->second.size());
          ++iMap;
       }  
    }
